% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gtsummary_core.R
\name{.create_gtsummary_object}
\alias{.create_gtsummary_object}
\title{Function uses \code{table_body} to create a gtsummary object}
\usage{
.create_gtsummary_object(table_body, ...)
}
\arguments{
\item{table_body}{the table_body tibble}

\item{...}{other objects that will be added to the gtsummary object list}
}
\value{
gtsummary object
}
\description{
Function uses \code{table_body} to create a gtsummary object
}
\keyword{internal}
