% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-as.R
\name{.table_styling_expr_to_row_number}
\alias{.table_styling_expr_to_row_number}
\title{Object Convert Helper}
\usage{
.table_styling_expr_to_row_number(x)
}
\arguments{
\item{x}{a gtsummary object}
}
\value{
a gtsummary object
}
\description{
Ahead of a gtsummary object being converted to an output type,
each logical expression saved in \code{x$table_styling} is converted to a
list of row numbers.
}
\examples{
tbl <-
  trial \%>\%
  tbl_summary(include = c(age, grade)) \%>\%
  .table_styling_expr_to_row_number()
}
\keyword{internal}
