% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_text.R
\name{inline_text.tbl_survival}
\alias{inline_text.tbl_survival}
\title{Report statistics from survival summary tables inline}
\usage{
\method{inline_text}{tbl_survival}(...)
}
\arguments{
\item{...}{Not used}
}
\description{
\lifecycle{deprecated}
}
\keyword{internal}
